requirements.txt
absl-py==1.4.0
ale-py==0.8.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brotlicffi==1.1.0.0
cachetools==5.3.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
cryptography==41.0.3
curio==1.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.8.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.4
fqdn==1.5.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.21.0
google-auth-oauthlib==1.1.0
grpcio==1.59.1
gviz-api==1.10.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.27.1
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.5.1
jupyter_server==2.12.4
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
notebook==7.2.0
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
path==16.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PySide6==6.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-socks==2.4.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
Send2Trash==1.8.2
sentry-sdk==2.2.0
setproctitle==1.3.3
shiboken6==6.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-profile==2.11.1
tensorboard-plugin-wit==1.7.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
torch==2.0.1
tornado==6.3.3
traitlets==5.10.0
trio==0.22.2
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wandb==0.17.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.8
zipp==3.16.2
